+++ SOOSA/SOOSA/src/InputConfiguration.cpp
@@ -24,8 +24,10 @@ string InputConfiguration::getPath() const { return m_path; }
 string InputConfiguration::getArea() const { return m_area; }
 string InputConfiguration::getPeriod() const { return m_period; }
 double InputConfiguration::getDischarge() const { return m_discharge; }
-int InputConfiguration::getNumberOfColumns() const { return m_formDetails.columnToNumberOfQuestionsMap.size(); }
-int InputConfiguration::getNumberOfQuestions() const { return m_formDetails.questions.size(); }
+int InputConfiguration::getNumberOfColumns() const {
+    return static_cast<int>(m_formDetails.columnToNumberOfQuestionsMap.size());
+}
+int InputConfiguration::getNumberOfQuestions() const { return static_cast<int>(m_formDetails.questions.size()); }
 
 int InputConfiguration::getNumberOfQuestionsAtColumn(int const column) const {
     int numberOfQuestions = 0;
